ae8593-conditional-message-crud/requirements.txt
# Add your Python dependencies here
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.5.2
